Address PR comments.
