+++ src/components/data/toolbar/Toolbar.js
@@ -6,7 +6,9 @@
  * data and also allows toggling the table or grid view.
  */
 
-import React, { useCallback, useEffect, useState } from "react";
+import React, { useState } from "react";
+
+import { injectIntl } from "react-intl";
 
 import DataDotMenu from "./DataDotMenu";
 import UploadMenuBtn from "./UploadMenuBtn";
@@ -18,32 +20,13 @@ import styles from "../styles";
 import CreateFolderDialog from "../CreateFolderDialog";
 import { isWritable } from "../utils";
 
-import UploadDialog from "../../uploads/dialog";
 import DisplayTypeSelector from "../../utils/DisplayTypeSelector";
 
-import { useUploadTrackingState } from "../../../contexts/uploadTracking";
+import { build, getMessage, withI18N } from "@cyverse-de/ui-lib";
 
-import { injectIntl } from "react-intl";
-
-import {
-    announce,
-    build,
-    formatMessage,
-    getMessage,
-    withI18N,
-} from "@cyverse-de/ui-lib";
-
-import {
-    Button,
-    Hidden,
-    makeStyles,
-    Toolbar,
-    Typography,
-    useTheme,
-} from "@material-ui/core";
+import { Button, Hidden, makeStyles, Toolbar } from "@material-ui/core";
 
 import { CreateNewFolder, Info } from "@material-ui/icons";
-import URLImportDialog from "../../URLImportDialog";
 
 const useStyles = makeStyles(styles);
 
@@ -63,48 +46,17 @@ function DataToolbar(props) {
         onDetailsSelected,
         handlePathChange,
         handleDataNavError,
-        intl,
+        setImportDialogOpen,
+        setUploadDialogOpen,
+        localUploadId,
+        uploadMenuId,
     } = props;
-    const theme = useTheme();
-    const tracker = useUploadTrackingState();
+
     const [createFolderDlgOpen, setCreateFolderDlgOpen] = useState(false);
-    const [uploadDialogOpen, setUploadDialogOpen] = useState(false);
-    const [importDialogOpen, setImportDialogOpen] = useState(false);
     const onCreateFolderDlgClose = () => setCreateFolderDlgOpen(false);
     const onCreateFolderClicked = () => setCreateFolderDlgOpen(true);
-    const onCloseImportDialog = () => setImportDialogOpen(false);
-
-    const viewUploadQueue = useCallback(() => {
-        return (
-            <Button
-                variant="outlined"
-                onClick={() => setUploadDialogOpen(true)}
-            >
-                <Typography
-                    variant="button"
-                    style={{ color: theme.palette.primary.contrastText }}
-                >
-                    {formatMessage(intl, "uploadQueue")}
-                </Typography>
-            </Button>
-        );
-    }, [intl, theme.palette.primary.contrastText]);
-
-    useEffect(() => {
-        if (tracker.uploads.length > 0) {
-            announce({
-                text: formatMessage(intl, "filesQueuedForUploadMsg", {
-                    total: tracker.uploads.length,
-                }),
-                CustomAction: viewUploadQueue,
-            });
-        }
-    }, [tracker, intl, viewUploadQueue]);
 
     let toolbarId = build(baseId, ids.TOOLBAR);
-    const uploadMenuId = build(toolbarId, ids.UPLOAD_MENU);
-    const localUploadId = build(toolbarId, ids.UPLOAD_MI, ids.UPLOAD_INPUT);
-
     return (
         <Toolbar variant="dense">
             <DisplayTypeSelector
@@ -180,15 +132,6 @@ function DataToolbar(props) {
                     onCreateFolderDlgClose();
                     refreshListing();
                 }}
-            />
-            <UploadDialog
-                open={uploadDialogOpen}
-                handleClose={() => setUploadDialogOpen(false)}
-            />
-            <URLImportDialog
-                path={path}
-                open={importDialogOpen}
-                onClose={onCloseImportDialog}
             />
         </Toolbar>
     );
